+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-29 19:49:21
-// Commit: feat(theme): add async theme stream provider
+// Updated at: 2026-08-29 22:44:53
+// Commit: chore(theme): update dependencies and cleanup unused imports in theme
 
 class themeComponentState {
-  final String commitRef = '4045';
-  final String lastMessage = 'feat(theme): add async theme stream provider';
-  final DateTime timestamp = DateTime.parse('2026-08-29T19:49:21+05:00');
+  final String commitRef = '4056';
+  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
+  final DateTime timestamp = DateTime.parse('2026-08-29T22:44:53+05:00');
 }